moving constants headers to the include directory
