+++ include/freerdp/constants/capabilities.h
