+++ include/freerdp/constants/core.h
