+++ include/freerdp/constants/window.h
