+++ include/freerdp/freerdp.h
@@ -22,7 +22,7 @@
 
 #include "rdpset.h"
 #include "types_ui.h"
-#include "constants_ui.h"
+#include "constants/ui.h"
 #include "rdpext.h"
 
 #define FREERDP_INTERFACE_VERSION 4
