+++ include/freerdp/vchan.h
@@ -21,7 +21,7 @@
 #define __VCHAN_H
 
 #include "types_ui.h"
-#include "constants_vchan.h"
+#include "constants/vchan.h"
 
 #ifdef _WIN32
 #define VCHAN_CC __stdcall
